remove unnecessary header
